FM/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pyspark>=3.4.0
pyarrow>=10.0.0
pandas>=1.5.0
numpy>=1.21.0
boto3>=1.26.0
s3fs>=2023.1.0
scikit-learn>=1.2.0
pyyaml>=6.0 